+++ hikyuu_cpp/hikyuu/trade_sys/moneymanager/crt/MM_FixedCount.h
@@ -21,7 +21,7 @@ namespace hku {
  *       2) 该策略并不判断已有的持仓情况，如果在已有持仓情况下不能交易，则该判断应为System本身的责任
  * @ingroup MoneyManager
  */
-MoneyManagerPtr HKU_API MM_FixedCount(int n = 100);
+MoneyManagerPtr HKU_API MM_FixedCount(double n = 100);
 
 }  // namespace hku
