+++ hikyuu_cpp/hikyuu/trade_sys/moneymanager/imp/FixedUnitsMoneyManager.cpp
@@ -15,7 +15,7 @@ FixedUnitsMoneyManager::FixedUnitsMoneyManager() : MoneyManagerBase("MM_FixedUni
 
 FixedUnitsMoneyManager::~FixedUnitsMoneyManager() {}
 
-size_t FixedUnitsMoneyManager ::_getBuyNumber(const Datetime& datetime, const Stock& stock,
+double FixedUnitsMoneyManager ::_getBuyNumber(const Datetime& datetime, const Stock& stock,
                                               price_t price, price_t risk, SystemPart from) {
     int n = getParam<int>("n");
     price_t fixed_risk =
